Move detection feasibility test to a non-auto-generated file
